backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Cache / Broker
redis[hiredis]==5.0.1
celery[redis]==5.3.6

# Config
pydantic-settings==2.1.0

# Auth
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0

# Validation
email-validator==2.1.0

# Logging
structlog==24.1.0

# Metrics (wired in Phase 11)
prometheus-client==0.20.0

# Testing
httpx==0.27.0
pytest==8.0.2
pytest-asyncio==0.23.5

# Linting / Typing
ruff==0.3.0
mypy==1.8.0